[RPC Framework] Add _script_module_reducer unconditionally for RecursiveScriptModule in RPC pickler
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58020

Previously there is no RPC pickler for `RecursiveScriptModule`. Although it is a subclass of `ScriptModule`, the reducer of `ScriptModule` is not triggered for `RecursiveScriptModule` when a script remote module is sent over RPC.

This PR checkpoints the investigation of #58274, which makes sure that a RPC pickler is invoked here. This still cannot fix `test_send_remote_module_over_the_wire_script`. Will revisit this bug once there is a feature request from users.

ghstack-source-id: 128949642

Differential Revision: [D28346758](https://our.internmc.facebook.com/intern/diff/D28346758/)